moved all callbacks into the message handler files and made callback registration fully automatic
