+++ ygo/message_handlers/announce_attrib.py
@@ -1,9 +1,18 @@
+import io
 from twisted.internet import reactor
 
 from ygo.constants import ATTRIBUTES
 from ygo.duel_reader import DuelReader
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel
+
+def msg_announce_attrib(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  count = self.read_u8(data)
+  avail = self.read_u32(data)
+  self.cm.call_callbacks('announce_attrib', player, count, avail)
+  return data.read()
 
 def announce_attrib(self, player, count, avail):
   attrmap = {k: (1<<i) for i, k in enumerate(ATTRIBUTES)}
@@ -32,3 +41,7 @@ def announce_attrib(self, player, count, avail):
     self.set_responsei(value)
     reactor.callLater(0, process_duel, self)
   return prompt()
+
+MESSAGES = {141: msg_announce_attrib}
+
+CALLBACKS = {'announce_attrib': announce_attrib}
