+++ ygo/message_handlers/announce_card.py
@@ -1,9 +1,17 @@
 from gsb.intercept import Reader
+import io
 from twisted.internet import reactor
 
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel
 from ygo import globals
+
+def msg_announce_card(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  type = self.read_u32(data)
+  self.cm.call_callbacks('announce_card', player, type)
+  return data.read()
 
 def announce_card(self, player, type):
   pl = self.players[player]
@@ -22,3 +30,7 @@ def announce_card(self, player, type):
     self.set_responsei(card.code)
     reactor.callLater(0, process_duel, self)
   prompt()
+
+MESSAGES = {142: msg_announce_card}
+
+CALLBACKS = {'announce_card': announce_card}
