+++ ygo/message_handlers/announce_card_filter.py
@@ -1,10 +1,21 @@
 from gsb.intercept import Reader
+import io
 from twisted.internet import reactor
 
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel
 from ygo import globals
 from ygo import duel
+
+def msg_announce_card_filter(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  size = self.read_u8(data)
+  options = []
+  for i in range(size):
+    options.append(self.read_u32(data))
+  self.cm.call_callbacks('announce_card_filter', player, options)
+  return data.read()
 
 def announce_card_filter(self, player, options):
   pl = self.players[player]
@@ -25,3 +36,7 @@ def announce_card_filter(self, player, options):
     self.set_responsei(card.code)
     reactor.callLater(0, process_duel, self)
   prompt()
+
+MESSAGES = {144: msg_announce_card_filter}
+
+CALLBACKS = {'announce_card_filter': announce_card_filter}
