+++ ygo/message_handlers/become_target.py
@@ -1,3 +1,16 @@
+import io
+
+def msg_become_target(self, data):
+  data = io.BytesIO(data[1:])
+  u = self.read_u8(data)
+  target = self.read_u32(data)
+  tc = target & 0xff
+  tl = (target >> 8) & 0xff
+  tseq = (target >> 16) & 0xff
+  tpos = (target >> 24) & 0xff
+  self.cm.call_callbacks('become_target', tc, tl, tseq)
+  return data.read()
+
 def become_target(self, tc, tl, tseq):
   card = self.get_card(tc, tl, tseq)
   if not card:
@@ -9,3 +22,7 @@ def become_target(self, tc, tl, tseq):
     if (pl.watching or card.controller != pl.duel_player) and card.position in (0x8, 0xa):
       tcname = pl._("%s card") % card.get_position(pl)
     pl.notify(pl._("%s targets %s (%s)") % (name, spec, tcname))
+
+MESSAGES = {83: msg_become_target}
+
+CALLBACKS = {'become_target': become_target}
