+++ ygo/message_handlers/draw.py
@@ -1,3 +1,19 @@
+import io
+
+from ygo.card import Card
+
+def msg_draw(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  drawed = self.read_u8(data)
+  cards = []
+  for i in range(drawed):
+    c = self.read_u32(data)
+    card = Card(c & 0x7fffffff)
+    cards.append(card)
+  self.cm.call_callbacks('draw', player, cards)
+  return data.read()
+
 def draw(self, player, cards):
   pl = self.players[player]
   pl.notify(pl._("Drew %d cards:") % len(cards))
@@ -7,3 +23,7 @@ def draw(self, player, cards):
   op.notify(op._("Opponent drew %d cards.") % len(cards))
   for w in self.watchers:
     w.notify(w._("%s drew %d cards.") % (pl.nickname, len(cards)))
+
+MESSAGES = {90: msg_draw}
+
+CALLBACKS = {'draw': draw}
