+++ ygo/message_handlers/hint.py
@@ -1,7 +1,16 @@
+import io
 from twisted.internet import reactor
 
 from ygo.utils import process_duel
 from ygo import globals
+
+def msg_hint(self, data):
+  data = io.BytesIO(data[1:])
+  msg = self.read_u8(data)
+  player = self.read_u8(data)
+  value = self.read_u32(data)
+  self.cm.call_callbacks('hint', msg, player, value)
+  return data.read()
 
 def hint(self, msg, player, data):
   pl = self.players[player]
@@ -13,3 +22,7 @@ def hint(self, msg, player, data):
   elif msg == 9:
     op.notify(globals.strings[op.language]['system'][1512] % data)
     reactor.callLater(0, process_duel, self)
+
+MESSAGES = {2: msg_hint}
+
+CALLBACKS = {'hint': hint}
