+++ ygo/message_handlers/pos_change.py
@@ -1,3 +1,19 @@
+import io
+
+from ygo.card import Card
+
+def msg_pos_change(self, data):
+  data = io.BytesIO(data[1:])
+  code = self.read_u32(data)
+  card = Card(code)
+  card.controller = self.read_u8(data)
+  card.location = self.read_u8(data)
+  card.sequence = self.read_u8(data)
+  prevpos = self.read_u8(data)
+  card.position = self.read_u8(data)
+  self.cm.call_callbacks('pos_change', card, prevpos)
+  return data.read()
+
 def pos_change(self, card, prevpos):
   cs = card.get_spec(card.controller)
   cso = card.get_spec(1 - card.controller)
@@ -8,3 +24,7 @@ def pos_change(self, card, prevpos):
   for w in self.watchers:
     cs = card.get_spec(w.duel_player)
     w.notify(w._("The position of card %s (%s) was changed to %s.") % (cs, card.get_name(w), card.get_position(w)))
+
+MESSAGES = {53: msg_pos_change}
+
+CALLBACKS = {'pos_change': pos_change}
