+++ ygo/message_handlers/retry.py
@@ -1,0 +1,5 @@
+def msg_retry(self, buf):
+  print("retry")
+  return buf[1:]
+
+MESSAGES = {1: msg_retry}
