+++ ygo/message_handlers/select_card.py
@@ -1,8 +1,47 @@
+import io
 from twisted.internet import reactor
 
+from ygo.card import Card
 from ygo.duel_reader import DuelReader
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel, parse_ints
+
+def msg_select_tribute(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  cancelable = self.read_u8(data)
+  min = self.read_u8(data)
+  max = self.read_u8(data)
+  size = self.read_u8(data)
+  cards = []
+  for i in range(size):
+    code = self.read_u32(data)
+    card = Card(code)
+    card.controller = self.read_u8(data)
+    card.location = self.read_u8(data)
+    card.sequence = self.read_u8(data)
+    card.position = self.get_card(card.controller, card.location, card.sequence).position
+    card.release_param = self.read_u8(data)
+    cards.append(card)
+  self.cm.call_callbacks('select_tribute', player, cancelable, min, max, cards)
+  return data.read()
+
+def msg_select_card(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  cancelable = self.read_u8(data)
+  min = self.read_u8(data)
+  max = self.read_u8(data)
+  size = self.read_u8(data)
+  cards = []
+  for i in range(size):
+    code = self.read_u32(data)
+    loc = self.read_u32(data)
+    card = Card(code)
+    card.set_location(loc)
+    cards.append(card)
+  self.cm.call_callbacks('select_card', player, cancelable, min, max, cards)
+  return data.read()
 
 def select_card(self, player, cancelable, min_cards, max_cards, cards, is_tribute=False):
   pl = self.players[player]
@@ -37,3 +76,11 @@ def select_card(self, player, cancelable, min_cards, max_cards, cards, is_tribut
     self.set_responseb(buf)
     reactor.callLater(0, process_duel, self)
   return prompt()
+
+def select_tribute(self, *args, **kwargs):
+  kwargs['is_tribute'] = True
+  self.select_tribute(*args, **kwargs)
+
+MESSAGES = {15: msg_select_card, 20: msg_select_tribute}
+
+CALLBACKS = {'select_card': select_card, 'select_tribute': select_tribute}
