+++ ygo/message_handlers/select_effectyn.py
@@ -1,7 +1,18 @@
+import io
 from twisted.internet import reactor
 
+from ygo.card import Card
 from ygo.parsers.yes_or_no_parser import yes_or_no_parser
 from ygo.utils import process_duel
+
+def msg_select_effectyn(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  card = Card(self.read_u32(data))
+  card.set_location(self.read_u32(data))
+  desc = self.read_u32(data)
+  self.cm.call_callbacks('select_effectyn', player, card, desc)
+  return data.read()
 
 def select_effectyn(self, player, card, desc):
   pl = self.players[player]
@@ -18,3 +29,7 @@ def select_effectyn(self, player, card, desc):
   if s != '':
     question += '\n'+s
   pl.notify(yes_or_no_parser, question, yes, no=no, restore_parser=old_parser)
+
+MESSAGES = {12: msg_select_effectyn}
+
+CALLBACKS = {'select_effectyn': select_effectyn}
