+++ ygo/message_handlers/select_option.py
@@ -1,11 +1,21 @@
 from gsb.intercept import Menu
-
+import io
 from twisted.internet import reactor
 
 from ygo.card import Card
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel
 from ygo import globals
+
+def msg_select_option(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  size = self.read_u8(data)
+  options = []
+  for i in range(size):
+    options.append(self.read_u32(data))
+  self.cm.call_callbacks("select_option", player, options)
+  return data.read()
 
 def select_option(self, player, options):
   pl = self.players[player]
@@ -25,3 +35,7 @@ def select_option(self, player, options):
   for idx, opt in enumerate(opts):
     m.item(opt)(lambda caller, idx=idx: select(caller, idx))
   pl.notify(m)
+
+MESSAGES = {14: msg_select_option}
+
+CALLBACKS = {'select_option': select_option}
