+++ ygo/message_handlers/select_place.py
@@ -1,8 +1,18 @@
+import io
 from twisted.internet import reactor
 
 from ygo.duel_reader import DuelReader
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel
+
+def msg_select_place(self, data):
+  data = io.BytesIO(data)
+  msg = self.read_u8(data)
+  player = self.read_u8(data)
+  count = self.read_u8(data)
+  flag = self.read_u32(data)
+  self.cm.call_callbacks('select_place', player, count, flag)
+  return data.read()
 
 def select_place(self, player, count, flag):
   pl = self.players[player]
@@ -34,3 +44,7 @@ def select_place(self, player, count, flag):
     self.set_responseb(resp)
     reactor.callLater(0, process_duel, self)
   pl.notify(DuelReader, r, no_abort=pl._("Invalid command"), restore_parser=DuelParser)
+
+MESSAGES = {18: msg_select_place, 24: msg_select_place}
+
+CALLBACKS = {'select_place': select_place}
