+++ ygo/message_handlers/select_position.py
@@ -1,8 +1,19 @@
 from gsb.intercept import Menu
+import io
 from twisted.internet import reactor
 
+from ygo.card import Card
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel
+
+def msg_select_position(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  code = self.read_u32(data)
+  card = Card(code)
+  positions = self.read_u8(data)
+  self.cm.call_callbacks('select_position', player, card, positions)
+  return data.read()
 
 def select_position(self, player, card, positions):
   pl = self.players[player]
@@ -19,3 +30,7 @@ def select_position(self, player, card, positions):
   if positions & 8:
     m.item(pl._("Face-down defense"))(lambda caller: set(caller, 8))
   pl.notify(m)
+
+MESSAGES = {19: msg_select_position}
+
+CALLBACKS = {'select_position': select_position}
