+++ ygo/message_handlers/sort_card.py
@@ -1,8 +1,24 @@
+import io
 from twisted.internet import reactor
 
+from ygo.card import Card
 from ygo.duel_reader import DuelReader
 from ygo.parsers.duel_parser import DuelParser
 from ygo.utils import process_duel, parse_ints
+
+def msg_sort_card(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  size = self.read_u8(data)
+  cards = []
+  for i in range(size):
+    card = Card(self.read_u32(data))
+    card.controller = self.read_u8(data)
+    card.location = self.read_u8(data)
+    card.sequence = self.read_u8(data)
+    cards.append(card)
+  self.cm.call_callbacks('sort_card', player, cards)
+  return data.read()
 
 def sort_card(self, player, cards):
   pl = self.players[player]
@@ -29,3 +45,7 @@ def sort_card(self, player, cards):
     self.set_responseb(bytes(ints))
     reactor.callLater(0, process_duel, self)
   prompt()
+
+MESSAGES = {25: msg_sort_card}
+
+CALLBACKS = {'sort_card': sort_card}
