+++ ygo/message_handlers/yesno.py
@@ -1,9 +1,17 @@
+import io
 from twisted.internet import reactor
 
 from ygo.card import Card
 from ygo.parsers.yes_or_no_parser import yes_or_no_parser
 from ygo.utils import process_duel
 from ygo import globals
+
+def msg_yesno(self, data):
+  data = io.BytesIO(data[1:])
+  player = self.read_u8(data)
+  desc = self.read_u32(data)
+  self.cm.call_callbacks('yesno', player, desc)
+  return data.read()
 
 def yesno(self, player, desc):
   pl = self.players[player]
@@ -21,3 +29,7 @@ def yesno(self, player, desc):
     opt = "String %d" % desc
     opt = globals.strings[pl.language]['system'].get(desc, opt)
   pl.notify(yes_or_no_parser, opt, yes, no=no, restore_parser=old_parser)
+
+MESSAGES = {13: msg_yesno}
+
+CALLBACKS = {'yesno': yesno}
